+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/validation/UIProcessValidator.scala
@@ -1,6 +1,6 @@
 package pl.touk.nussknacker.ui.validation
 
-import cats.data.NonEmptyList
+import cats.data.{NonEmptyList, Validated}
 import cats.data.Validated.{Invalid, Valid}
 import pl.touk.nussknacker.engine.api.component.ScenarioPropertyConfig
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError
@@ -21,6 +21,7 @@ import pl.touk.nussknacker.restmodel.validation.ValidationResults.{
 }
 import pl.touk.nussknacker.ui.definition.{DefinitionsService, ScenarioPropertiesConfigFinalizer}
 import pl.touk.nussknacker.ui.process.fragment.FragmentResolver
+import pl.touk.nussknacker.ui.process.label.ScenarioLabel
 import pl.touk.nussknacker.ui.process.marshall.CanonicalProcessConverter
 import pl.touk.nussknacker.ui.security.api.LoggedUser
 
@@ -29,6 +30,7 @@ class UIProcessValidator(
     validator: ProcessValidator,
     scenarioProperties: Map[String, ScenarioPropertyConfig],
     scenarioPropertiesConfigFinalizer: ScenarioPropertiesConfigFinalizer,
+    scenarioLabelsValidator: ScenarioLabelsValidator,
     additionalValidators: List[CustomProcessValidator],
     fragmentResolver: FragmentResolver,
 ) {
@@ -44,6 +46,7 @@ class UIProcessValidator(
       validator,
       scenarioProperties,
       scenarioPropertiesConfigFinalizer,
+      scenarioLabelsValidator,
       additionalValidators,
       fragmentResolver
     )
@@ -54,25 +57,20 @@ class UIProcessValidator(
       transform(validator),
       scenarioProperties,
       scenarioPropertiesConfigFinalizer,
+      scenarioLabelsValidator,
       additionalValidators,
       fragmentResolver
     )
 
-  // TODO: It is used only in tests, remove it from the prodcution code
-  def withScenarioPropertiesConfig(scenarioPropertiesConfig: Map[String, ScenarioPropertyConfig]) =
-    new UIProcessValidator(
-      processingType,
-      validator,
-      scenarioPropertiesConfig,
-      scenarioPropertiesConfigFinalizer,
-      additionalValidators,
-      fragmentResolver
-    )
-
-  def validate(scenarioGraph: ScenarioGraph, processName: ProcessName, isFragment: Boolean)(
+  def validate(
+      scenarioGraph: ScenarioGraph,
+      processName: ProcessName,
+      isFragment: Boolean,
+      labels: List[ScenarioLabel]
+  )(
       implicit loggedUser: LoggedUser
   ): ValidationResult = {
-    val uiValidationResult = uiValidation(scenarioGraph, processName, isFragment)
+    val uiValidationResult = uiValidation(scenarioGraph, processName, isFragment, labels)
 
     // TODO: Enable further validation when save is not allowed
     // The problem preventing further validation is that loose nodes and their children are skipped during conversion
@@ -91,8 +89,14 @@ class UIProcessValidator(
   // is an error preventing graph canonization. For example we want to display node and scenario id errors for scenarios
   // that have loose nodes. If you want to achieve this result, you need to add these validations here and deduplicate
   // resulting errors later.
-  def uiValidation(scenarioGraph: ScenarioGraph, processName: ProcessName, isFragment: Boolean): ValidationResult = {
+  def uiValidation(
+      scenarioGraph: ScenarioGraph,
+      processName: ProcessName,
+      isFragment: Boolean,
+      labels: List[ScenarioLabel]
+  ): ValidationResult = {
     validateScenarioName(processName, isFragment)
+      .add(validateScenarioLabels(labels))
       .add(validateNodesId(scenarioGraph))
       .add(validateDuplicates(scenarioGraph))
       .add(validateLooseNodes(scenarioGraph))
@@ -180,6 +184,23 @@ class UIProcessValidator(
     nodeIdErrors match {
       case Some(value) => formatErrors(value)
       case None        => ValidationResult.success
+    }
+  }
+
+  private def validateScenarioLabels(labels: List[ScenarioLabel]): ValidationResult = {
+    scenarioLabelsValidator.validate(labels) match {
+      case Valid(()) =>
+        ValidationResult.success
+      case Invalid(errors) =>
+        ValidationResult.globalErrors(
+          errors
+            .map(ve =>
+              ScenarioLabelValidationError(label = ve.label, description = ve.validationMessages.toList.mkString(", "))
+            )
+            .map(PrettyValidationErrors.formatErrorMessage)
+            .map(UIGlobalError(_, nodeIds = List.empty))
+            .toList
+        )
     }
   }